+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2021-3-18
+## [Unreleased] - 2021-3-19
 
 ### Added
 * Added an implementation of the TwoMax problem.
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   search landscape with one global optima, and one sub-optimal local
   optima, where most of the search space is within the attraction basin 
   of the local optima.
+* Added an implementation of Ackley's Porcupine function, an artificial
+  search landscape with one global optima, and an exponential number of
+  local optima, a very rugged landscape.
 
 ### Changed